Comments, style and transformers.Templater
The Templater transformer is an example of a simple transformer, and
allows us to remove templater-logic from the remaining code base. It's
not perfect (mainly because it wastes memory and causes a shit-load of
copy/writes), but the simplicity it provides probably justifies the
minor performance hit on machines that can do multiple GB's of memory
reads and writes.