Introduce Parsers to Handlers to decouple transport from container
I'm planning to shortly write a "JSON-parsed BMP" parser, a M&R
"port collector" parser (and Sender), a "TCP line receiver" and a
FIFO line receiver.

The idea is simple: There are multiple tools that can send HTTP data,
but not all of them will use the same format for the data. Similarily,
several tools are able to transmit data both over HTTP and other
formats.

This decoupling makes this very easy.