Add skogul-x2y, first go at generic receiver/sender
Future variants will not need to iterate over all possible senders and
receivers, but instead move the responsibility to the senders themself,
as suggested in the comment.

In the meanwhile, this little thing provides a good start at reducing
the need for X tiny binaries that basically all do the same thing.