Write http benchmarks and fix keep-alive sillyness
Prior to this commit, we did not correctly drain the body of a response
in the http sender, this prevented the HTTP client from correctly
identifying the connection as idle, and as such also prevented
connection re-use. Instead, connections would accumulate until they
timed out.

This was revealed by the benchmarking. I also switched from 200 to 204
in the response message, which circumvents the entire issue anyway,
since a 204 is OK, but without a response body.

An other issue is the default MaxIdleConnsPerHost, which is set to 2.
This means that for any setup where more than two go routines want to
push data to the same host, they will frequently end up creating new
connections and then closing them again. This both accumulates TIME_WAIT
connections, and increases a silly overhead in creating new connections
all the time.