Add Influxdb MeasurementFromMetadata setting
Fixes #12

The big XXX here isn't really new to influxdb. I'm thinking a log
message is in order, but I've left it out for now because I really think
we need to revamp the logging in general now that I've/we've got a
better understanding of what we really need.