http-sender: Support alternate root CA
Fixes #43

Had to use ht.ok, since init() is just run once. It shouldn't fail,
since Verify will check if it works, but in the scenario where something
changes between startup/Verify and the first metric sent with the
sender, it could potentially fail, so we need to handle it.

Not really any way to prevent that without an API change.