Generalize modules and expose custom data structures
The worthwhile changes here moves most of the logic from */auto.go into
modules.go. It's the same basic logic, but instead of
Receiver/Sender/Transformer, there's a generic Module structure.

This dramatically reduces the complexity of config/help.go and
cmd/skogul/main.go.

This also introduces "Extras" to the Module data structure, which allows
modules to list a set of custom data structures that should also be
visible in documentation. The Extras field is currently _only_ used for
documentation.

For review, begin with modules.go, comparing it to e.g. sender/auto.go.

Fixes #73