Address a nasty reference issue with TransformerMap
So here's the story: A []TransformerRef will be an array of references
to transformers. Ok, all good so far.

The problem was this:

Step 1: The TransformerRef is unmarshalled, causing a *pointer* to it to
        be put on the TransformerMap.
Step 2: The array is appended to. But it needs to be grown. This is done
        by *copying* the content. Suddenly, the array contain a copy of
	the original TransformerRef, but TransformerMap points to the
	original.
Step 3: resolveTransformer() iterates over the TransformerMap, which has
        no knowledge of the copy of the TransformerRef.
Step 4: resolveHandlers() uses the pre-resolving reference.

This is actually a big problem. We can't track the content of a
ResolverRef, only its creation thorugh Unmarshal(). We _might_ solve
this by something along the lines of making TransformerRef.T be a
pointer to a Transformer, and then have the TransformerMap track that
instead.

For now, this is a correct solution, but it doesn't solve the issue of
extremely subtle bugs for module developers who want to do the same
thing.... which reminds me I need to go over any code that uses
[]somethingRef....