split: Handle invalid data better (or at all)
Fixes #102

The updated test case didn't work prior to this, which demonstrates the
issue well: With a container of mixed metrics, where all (or several)
metrics have the same field that the split transformer wants to split
on, but it isn't actually "split-able", the transformer would just stop.

This ensures the transformer obeys Failed == false setting, and silently
continues, leaving "incompatible" containers intact.