Promote parsers to fully fledged modules
This ALSO introduces the concept of "automake" modules, which are
modules that do not need to be explicitly defined in their relevant
configuration section. This will let us continue to refer to parsers as
just "json" without having to explicitly define the json-parser, which
doesn't require any options, but gives us the option to do so.

In addition to most parsers, this has been enabled for the debug-sender
and the templater transformer.

The rest of this patch is mostly just a matter of replacing the
hardcoded resolveParser() with the resolveParsers() function that is
similar to resolveSenders, etc, and the required scaffolding to support
that with various Marshal/Unmarshal functions.

All tests pass.

NOTE: This introduces one potentially breaking change: In the past, you
could omit the parser from the handler section and it would silently
default to json. This is no longer allowed and a parser MUST be
explicitly configured for every handler. Now that we have a decent
amount of parsers, this seems justified.