Add "default" name to the debug-sender and add library
This is the first of hopefully many small steps to streamline using
skogul for complex setups.

The first part is very minor: Add an alias for the debug-sender called
"default" which is meant to be overridden by users. The second part adds
the beginning of a library scheme for pre-made configuration files.
Specifically for using Junos Telemetry. The library (in lack of a better
name) is what we use for receiving junos telemetry, but sends to the
"default" sender, so can be applied for anyone.

The general scheme probably needs to be adjusted as more libraries are
added.