main: Tweak the -help output to list all modules
This also fixes how aliases are listed. I'm not sure this output SHOULD
be part of -help, but I find it quite useful myself for quick reference.
At any rate, this commit makes it more accurate, and if we later decide
to split it into a separate function that's a separate discussion
regardless.