Add somewhat ugly support for splitting on dict
The split transformer can already split on arrays, and the exact same
code can theoretically do the same thing on any type of object you can
run "range" on. However, it needs to be cast.

This PR is meant as a WIP, since the implementation is not particularly
good since I couldn't find a generic way of handling this, but instead
just duplicated the code, verbatim.