Add DictSplit transformer and expand Split somewhat
The code is depressingly duplicated, save for the casting, but I
honestly don't know how to remedy that.

This commit does TWO things:

1. It introduces a DictSplit transformer, which is the
   map[string]interface{} equivalent to the []interface{} version which
   is Split.
2. It adds MetadataName to both of them, which allows saving the key
   which is split on to metadata. For arrays, this will be the array
   index, starting with 0. For dict, it's the key of the dictionary,
   which is clearly useful.

There's a SMALL catch with the test cases here... They can, given "bad
luck" fail. This is because map[string]interface{} is unsorted, and we
assume that it's parsed in the order in which it is written. I could
generalize this, but it's a heck of a lot of work compared to just going
"oh well" and running the test again? We'll see, I suppose?