feat(influxdb): Add an experimental feature for converting metadata keys from float to string in a more sane way
The default in Go is to convert floats to fixed-width strings
of the numbers using scientific notation, but trying to do a
lookup using that value afterwards is pretty horrific.

Notes can be taken for data sources to not send
identifiers as numerical values and rather send them as strings
directly.

This feature is experimental and will *likely* be changed before it
lands in Skogul.