Start work on improving error reporting
This is the bare beginnings on a revamp on error handling. It's been
messy for a while and still is. The general idea is:

1.

Modules EITHER log OR return an error, ideally never both, with the only
exception being when it's reasonable to add extensive debug-info on
trace/debug.

2.

Do some tests with skogul.Identity[]. I am just not sure what the future
of skogul.Error is and I want to build some real world experience before
I do anything to core to automate it, thus this somewhat cumbersome
approach for now.

3.

Wrap errors. This is slightly complicated because it doesn't really work
well with logrus imo. E.g.: A module might have an error which
identifies itself, but by the time it is logged, that's just text. I
would rather have trivial error handling with fmt.Errorf("foo: %w", err)
than contrived logrus formats though, so we'll see. Similar to point 2,
I want more real word experience before these ideas are implemented all
over.