Add Cast transformer
This should address #195 partially at least. It allows an operator to
specify an admittedly shallow schema to enforce data types. It does not,
however, attempt to solve the general problem.

I still say:

Fixes #195

Because #195 should be addressed as explicitly as possible, e.g.: If a
data provider some times sends text and some times integers, the
operator should FIX that, and we shouldn't hide it. But failing that, we
should provide tools for the operator to - knowingly - work around bad
providers. This fix does that for the vast majority of issues we've seen
like this.

It will still require fixes like #218, which is a much deeper issue.