Add IgnorePartialErrors and Describe()
This is not quite done, and I'm sorry for the double-commit if anyone
sees this.

Describe() is used to, well, describe a metric and/or a Container. It is
meant to improve debug output, and standardize how a metric or container
is described. The end goal is to replace "Something went wrong, this is
what went wrong" with "Something went wrong with THIS TYPE OF DATA, this
is what went wrong.". This is particularly important in complex setups
where you have large streams of data where most of the data seems to
work fine, but not everything.

I will be improving this as I go along and use it to fix a real world
issue.

The second part adds IgnorePartialErrors, see #220. This is not really
complete, it just ignores the errors right now, but still pass them
along to senders, which is just wrong on so many levels.