Enrichment improvements, and a quicky to add profiling
-pprof can now be used to set up a http endpoint to serve profiling
data, used by 'go tool  pprof -web'.

The tweaks to enrichment are mostly realization that we don't need to
invent our own hash (which was pretty bad anyway). It also changes
loading to be sequential instead of in bulk, which lower RAM usage
somewhat, but not nearly as much as it should.