Remove superfluous record-keeping on influx parsing
Fixes #230

I have no idea how this was supposed to have worked to be honest, but
the test cases still pass, and it doesn't crash. In short, the "start"
variable tracks where the _next_ item starts, but the moment we start
modifying the data-array, start refers to an index out of bounds. Since
we aren't actually using it for anything but tracking where the data
ends, and we end up copying the data multiple times anyway, let's just
remove all that record-keeping and avoid bugs.