Add dummy-save parser
The dummy save parser will eat any byte array and store it to a file for
debugging, then just return a bogus container. The purpose is when you
have a data stream we're about to write a parser for, but we need to
grab example packets, so instead of writing a separate tool to connect
to Kafka, MQTT, HTTP or whatever, this will just solve that for you with
the same settings we're used to.