Remove a large amount of redundant log messages
This is mostly a matter of removing variants of:

	somethingLog.WithError(err).Error(...)
	return err

This almost always leads to (at least) duplicated log messages. The new
regime is: You EITHER return an error OR log an error, never both.

Some minor log-tweaks in the same general neighbourhood too, to simplify
the language. Some messages were saying: "file BLATTI failed to open:
(err)", but the actual (err) repeated the text almost verbatim leading
to unhelpfully long error messages.

I do need a review for some of this though...