Fix logrus+linefile test issue
The core issue was that the logrus test failed if the new
linefileadvanced tests also ran, but neither failed isolated.

This was caused by the linefile advanced test using a delay of 1s, which
which means the linefile advanced receiver keeps running - AND LOGGING -
every second, adding unexpected log entries for the logrus test case.